Refactor to isolate orientdb code (#17)
* Refactor orient code

* capture IT startup log, debug

* add explicit geronimo-atinject_1.0_spec
