+++ .circleci/circleci-readme.md
@@ -11,7 +11,7 @@ The local build runs in a docker container.
 
   * Run a local build with the following command:
           
-        circleci local execute -c .circleci/local-config.yml --job 'github-maven-deploy/build-and-test'
+        circleci local execute -c .circleci/local-config.yml --job 'build_and_test'
 
     With the above command, operations that cannot occur during a local build will show an error like this:
