+++ nexus-repository-cpan/src/main/java/org/sonatype/nexus/repository/cpan/orient/package-info.java
@@ -10,30 +10,9 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.repository.cpan.internal;
+@FeatureFlag(name = ORIENT_ENABLED)
+package org.sonatype.nexus.repository.cpan.orient;
 
-import javax.annotation.Nonnull;
+import org.sonatype.nexus.common.app.FeatureFlag;
 
-import org.sonatype.nexus.repository.cache.CacheControllerHolder;
-import org.sonatype.nexus.repository.cache.CacheControllerHolder.CacheType;
-
-/**
- * Asset kinds for CPAN.
- */
-public enum AssetKind
-{
-  ARCHIVE(CacheControllerHolder.CONTENT),
-  VARIOUS(CacheControllerHolder.METADATA),
-  CHECKSUM(CacheControllerHolder.METADATA);
-
-  private final CacheType cacheType;
-
-  AssetKind(final CacheType cacheType) {
-    this.cacheType = cacheType;
-  }
-
-  @Nonnull
-  public CacheType getCacheType() {
-    return cacheType;
-  }
-}
+import static org.sonatype.nexus.common.app.FeatureFlags.ORIENT_ENABLED;
