+++ nexus-repository-cpan/src/test/java/org/sonatype/nexus/repository/cpan/internal/CpanDataAccessTest.java
@@ -16,6 +16,7 @@ import java.util.List;
 
 import org.sonatype.goodies.testsupport.TestSupport;
 import org.sonatype.nexus.repository.Repository;
+import org.sonatype.nexus.repository.cpan.orient.internal.CpanDataAccess;
 import org.sonatype.nexus.repository.storage.Asset;
 import org.sonatype.nexus.repository.storage.Bucket;
 import org.sonatype.nexus.repository.storage.Component;
